+++ CHANGES.rst
@@ -1,3 +1,9 @@
+0.14.0 (2021-09-30)
+-------------------
+- One can now use python -m docxtpl on command line
+  to generate a docx from a template and a json file as a context
+  Thanks to Lcrs123@github
+
 0.12.0 (2021-08-15)
 -------------------
 - Code has be split into many files for better readability